# PERF: EWMA with times