--- a/pandas/core/window/ewm.py
+++ b/pandas/core/window/ewm.py
@@ -335,20 +335,22 @@ class ExponentialMovingWindow(BaseWindow):
     def mean(self, *args, **kwargs):
         nv.validate_window_func("mean", args, kwargs)
         if self.times is not None:
-            window_func = window_aggregations.ewma_time
-            window_func = partial(
-                window_func,
-                times=self.times,
-                halflife=self.halflife,
-            )
+            com = 1
+            times = self.times.astype(np.float64)
+            halflife = float(self.halflife)
         else:
-            window_func = window_aggregations.ewma
-            window_func = partial(
-                window_func,
-                com=self.com,
-                adjust=self.adjust,
-                ignore_na=self.ignore_na,
-            )
+            com = self.com
+            times = np.arange(len(self.obj), dtype=np.float64)
+            halflife = 1.0
+        window_func = window_aggregations.ewma
+        window_func = partial(
+            window_func,
+            com=com,
+            adjust=self.adjust,
+            ignore_na=self.ignore_na,
+            times=times,
+            halflife=halflife,
+        )
         return self._apply(window_func)
 
     @doc(