- [x] closes #39784
- [x] tests added / passed
- [x] Ensure all linting tests pass, see [here](https://pandas.pydata.org/pandas-docs/dev/development/contributing.html#code-standards) for how to run them
- [x] whatsnew entry

```
In [1]: idx=pd.date_range('20000101','20201231',periods=50000)
   ...:
   ...: df=pd.DataFrame(data=range(50000),index=idx)

In [2]: %timeit df.ewm(halflife=pd.Timedelta('100d'),times=df.index).mean()

1.2 ms ± 13.8 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) <-- PR

10.2 s ± 32.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each) <-- master
```