@jorisvandenbossche 2021-02-28
This seems to have resulted in a performance regression for the existing ASV benchmark: https://pandas.pydata.org/speed/pandas/#rolling.EWMMethods.time_ewm?python=3.8&Cython=0.29.21&p-constructor='DataFrame'&p-constructor='Series'&p-window=1000&p-dtype='float'&p-method='mean'&commits=ab687aec-f066cda6